+++ datanode/api/errors.go
@@ -324,6 +324,8 @@ var (
 
 	// Paid liquidity fees.
 	ErrListPaidLiquidityFees = errors.New("failed to list paid liquidity fees")
+	// List Games.
+	ErrListGames = errors.New("failed to list games")
 )
 
 // errorMap contains a mapping between errors and Vega numeric error codes.